requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.20.1
bottle==0.12.21
bottle-websocket==0.2.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
commonutil==0.1.20
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
image-quality==1.2.7
imageio==2.19.3
joblib==1.1.0
libsvm==3.23.0.4
macholib==1.16
networkx==2.8.3
numpy==1.22.4
olefile==0.46
opencv-python==4.6.0.66
packaging==21.3
pefile==2022.5.30
Pillow==9.1.1
pybrisque==1.0
pycparser==2.21
pyparsing==3.0.9
pyside2==5.15.2.1
PyWavelets==1.3.0
pywin32==303
pywin32-ctypes==0.2.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
tifffile==2022.5.4
urllib3==1.26.9
whichcraft==0.6.1
wincertstore==0.2
xmltodict==0.13.0
zope.event==4.5.0
zope.interface==5.4.0
